Support fieldsets in authentication settings forms.
The authentication settings forms had to be simple before, only
displaying the fields in the form in the order listed.

This change takes advantage of a new template in Djblets, allowing for
the use of fieldsets in auth settings forms.

Testing Done:
Checked that all current auth backends look as they did before.

Made use of fieldsets on another auth backend, and saw that the
fieldsets were used.

Reviewed at https://reviews.reviewboard.org/r/8375/