Fix onclick bug triggering invalid diff modes
Previously, it was possible to click between two image diff modes and
invoke the image mode changed function (due to padding). Now, the class
is directly on the `<a>` tag instead of the surrounding `<li>` tag,
which prevents this issue. This was causing an error (due to a missing
`data-mode` attribute on the `<li>` (the event target) that caused the
image review UI to stop functioning.

Testing done:
Clicked between two image modes. The image review UI no longer breaks.