Simplify review thread markup and CSS.
This change dramatically improves the markup and CSS for review threads.
Specifically, instead of using `<dl>` for everything (which isn't semantically
correct), we wrap things in `<ol>` elements, and then individual comments are
broken up into `<div>`s with sensible classes. These are then styled with
consistent and simple spacing so that everything looks nice while also being
aligned. Replies to a comment are all indented a bit from that comment,
emphasizing the relationship in the way that most threaded displays work. One
particularly nice part of this change is that the text for comments is now
horizontally aligned with the avatar (the borders of the text area are
accounted for using negative margins).

Testing done:
Tested reviews with all sorts of different comments, and replies to those
reviews. Checked to make sure everything was aligned and spaced correctly.
Opened the editor for draft replies to verify that there were no jumps in the
position of the text.

Reviewed at https://reviews.reviewboard.org/r/8391/