Status Updates part 4: Add a Feature definition.
Status updates are a new feature which are designed for better integration with
third-party tools like static analysis and continuous integration. This change
adds a Feature definition which will be used to gate off access to the new code
until it's ready to go.

This is not yet used by anything.

Testing done:
Used in future changes.

Reviewed at https://reviews.reviewboard.org/r/8405/