Backport support for Djblets' EmailMessage to Review Board 2.5.x.
Back in the Review Board 2.5 development cycle, the old
`"SpiffyEmailMessage` class was split out into an `EmailMessage` in
Djblets, allowing us to use the same logic in other products. We then
later updated Review Board 3.0 to make use of this class, but we left
the version in 2.5 alone, given where we were in the release cycle.

Now that the class has been well-tested, and has functionality we want
in Review Board 2.5, we can safely migrate over to it. This change
backports the Review Board 3.0 change to use the Djblets `EmailMessage`. It
also updates newer code to use it, and updates all call sites to pass
keyword arguments instead of positional (given the addition of newer fields
in the class).

This will be needed before we can begin using DMARC support in Review Board.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8414/