Move image scale into the ImageReviewableModel
The image scale is now stored in the `RB.ImageReviewableModel` instead
of being a property of the views. This method allows us to listen for
changes to the image scale, as well as set it programmatically and have
the UI update.

Testing Done:
- Ran JS tests.
- Ensured that changing the image scale still worked and the UI still
  updated. Comment boxes still scale appropriately.

Reviewed at https://reviews.reviewboard.org/r/8400/