Scroll oversized images in the image review UI
Previously, oversized images in the image review UI would cause them to
overflow outside of the review box and increase the width of the page.
Now, these images are confined to the width of the review box and will
scroll horizontally when they are too large.

Testing Done:
- Created a review comment on an oversized image; it scrolled with the
  page.
- Scrolled an oversized image and created a commment; it worked
  correctly and scrolled with the page.
- Ensured the scroll bar only showed when images were oversized for the
  currently selected scale.

Reviewed at https://reviews.reviewboard.org/r/8377/