Auto scale large images according to size
Previously, we scaled images that contained `@2x` or `@3x` in their
filenames. For images that do not contain those hints, we now scale them
to the highest scaling factor less than or equal to 1 that allows the
most of the image to be shown.

Testing Done:
Tested with a very large image and saw that it was scaled correctly in
image diff and single image mode.

Reviewed at https://reviews.reviewboard.org/r/8392/