Fix the attributes of the resolution menu
During review, the code for generating the resolution menu changed from
using a `for [a, b] of scalingFactors.items()` loop to a
`scalingFactors.foreach((a, b))` call. However, the parameters for the
callback to `Map.prototype.foreach` are backwards compared to what is
returned by `Map.prototype.items`. This lead to the resolution menu
having its text and `data-image-scale` attribute swapped. This patch
addresses that issue.

Testing Done:
- Verified that the text and data attributes were generated correctly.
- Verified that the menu switched the resolution.

Reviewed at https://reviews.reviewboard.org/r/8427/