Add tests for using GroupForm and DefaultReviewerForm with positional args.
The regression in the GroupForm was caused by the fact that the admin site
passes in the data to bind to the form as a positional argument rather than a
keyword argument. This meant that the data was being stuck into the
`local_site_name` argument rather than getting passed through to the base
class.

This change adds some unit tests to make sure we don't cause a similar problem
in the future. I also cleaned up some of the docstrings for adjacent tests to
match our conventions.

Testing done:
Ran the new tests. Before the fix to the forms, the new tests fail. After, they
run successfully.

Reviewed at https://reviews.reviewboard.org/r/8431/