Make sure test runner is imported after setting environment up.
We were importing the extension test runner before setting the
RB_RUNNING_TESTS environment variable which meant that the instance of
Review Board used for testing would have the wrong configuration
to start. Specifically, the `RUNNING_TESTS` setting was set to the
wrong value, causing parts of the codebase to be unaware that they
were running in a test environment, which caused them to set up
some defaults in the database that conflicted with the test fixtures.

Testing Done:
Wrote a simple test for an extension and ran it successfully.

Reviewed at https://reviews.reviewboard.org/r/8435/