Clean up some old cruft from SCMTools.
This removes an old method (`get_fields()`), which is no longer used in
any way. This used to be used to influence the New Review Request
dialog, but that was rewritten long ago.

It also replaces `get_diffs_use_absolute_diffs()` with a
`diffs_use_absolute_diffs` attribute (with defaults to being a property
that supports the old method for backward-compatibility).

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8428/