Status Updates part 9: centralize code to expand boxes to initial hash.
When loading a page that has something like "#review20" in the window location,
we had some code to expand the review box that matched. The addition of reviews
within the initial status reports and change description boxes duplicated this
code two additional times.

This change moves that responsibility into the `ReviewRequestPageView`. This
now just generically searches each box for a matching selector, expanding if it
finds it. This means that it's much more general, and linking to individual
comments will also expand their container.

Testing done:
Checked that the matching boxes were correctly expanded when loading the page
with a variety of different hash values.

Reviewed at https://reviews.reviewboard.org/r/8446/