Status Updates part 10: Make reply draft banners work inline in status updates.
Until now, my plan was to get as much of the status update code in as possible
but never move the future out of `EXPERIMENTAL` until we had the new unified
draft banners. This was because the reply banners for replies to reviews were
owned by the `ReviewBoxView`, and there seemed to be no good way to port that
behavior over given that there are multiple reviews stuffed into a single box.

This change makes it work by displaying the reply draft banners inline in the
status update sections, below the header (which contains the status update
summary) and above all the comments. This is set to float so that it will dock
at the top of the screen as long as that review is visible, the way that our
other reply draft banners work.

The goal is still to replace these with the new unified draft banners, but that
work is no longer blocking.

While I was in here I fixed a small regression that I introduced with
collapsing review boxes that have a draft reply (where the banner would still
be displayed).

Testing done:
- Replied to reviews attached to status updates, both in the initial status
  updates box and the change description boxes.
- Verified the floating banner behavior when scrolling around with a bunch of
  reply drafts.
- Checked the appearance of all types of boxes when expanding and collapsing.

Reviewed at https://reviews.reviewboard.org/r/8447/