Fix the links generated by the ReviewGroupUser resource
Previously the links generated by the `ReviewGroupUser` resource
pointed at the `UserResource`. In other words, the delete link for the
`ReviewGroupUser` resource would actually be the delete link for the
`UserResource` and using it would result in the deletion of the user,
not the removal from review group. This was because these resources
shared a name (`user`) and as such, the lookups for URL names were
incorrect.

Now we correctly generate the links for the `ReviewGroupUser` resource.
We have added a related link to the actual `user` resource. Unit tests
have been added which ensure the delete links are correct in the
resources.

Testing Done:
Ran unit tests.

Bugs closed: 4382

Reviewed at https://reviews.reviewboard.org/r/8449/