Fix post-commit support for Bitbucket.
The Bitbucket post-commit support was failing to generate a suitable
diff for display. When fetching a commit, the `base_commit_id` field was
being set to the ID of the commit, rather than the parent commit. This
led to fetching the patched version of the files for the diff, rather
than the source files.

The `Commit.base_commit_id` field is deprecated and redundant these days.
We instead want to use `Commit.parent`. This change updates everything to
use that, and passes it through as the base commit ID for file
fetches.

Deprecation docs have also been updated accordingly.

Testing Done:
Unit tests pass.

Manually tested post-commit on Bitbucket and GitLab (the two that both
support post-commit _and_ use `base_commit_id`.

Reviewed at https://reviews.reviewboard.org/r/8460/