Support users in multiple review groups in API responses
The patch that fixed the review group user resource links broke when a
user was a member of more than one group (which is a likely occurance).
This was because we were using a `model_parent_key` that was a many-to-
many field (and therefore the reverse relation was not unique),
resulting in a `MultipleObjectsReturned` exception. This has been fixed
by customizing how we retrieve the parent object IDs.

Testing Done:
Ran unit tests.

Bugs closed: 4382

Reviewed at https://reviews.reviewboard.org/r/8458/