Improve styling for the auth forms for errors, mobile, and number fields.
This fixes up some small issues with styling on authentication forms
(login, register, etc.), particularly for RBCommons.

On mobile, the form is no longer slightly off-center. There was an
unwanted margin on the right that shouldn't have been there, and was
noticeable given the screen size.

Form errors were appearing off-center, instead of aligned with the
field. They're now lined up properly.

Number fields (which are used on RBCommons, but not Review Board) didn't
have the same styling as the other input fields. This adds the same
styling for those fields.

<button> elements were also failing to be styled like other buttons.
This is also fixed.

Testing Done:
Tested the mobile and desktop versions of these styles on Review Board
and RBCommons.

Reviewed at https://reviews.reviewboard.org/r/8456/