Fix API doc building for review group user resources.
The review group user resource added a requirement to a 'group_name'
argument in the URL, in order to build links. API doc building doesn't
really know how to deal with this (due to the possibility of a user
having multiple parents, and not using the parent resource support).

Since we don't have a mechanism to provide this information, we're going
to just turn off the links reference for this resource's docs. Examples
will still work and will show the appropriate details.

Testing Done:
Built the API docs successfully.

Checked the resulting docs for this resource and saw that the links
reference was gone, but the examples were fine.

Reviewed at https://reviews.reviewboard.org/r/8468/