Allow Features to disable web API resources
Review Board web API resources now have a `required_features` attribute
which indicates which features **must** be enabled for the feature to be
available. If any of the listed features are disabled, the resource will
return a 403 Forbidden error for any of the resource's methods. When all
features are enabled, the resource will behave as normal.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8470/