Add a general comments feature class
The new general comments feature can be enabled or disabled with the
`reviews.general_comments` feature. Disabling this feature will disable
all UI elements for it, as well as the web API resources for it.

Testing Done:
Tested with `reviews.general_comments` enabled and disabled. When
disabled, general comments cannot be created and the UI elements do not
appear.

Reviewed at https://reviews.reviewboard.org/r/8471/