Require the status updates feature for the status updates resource
The status updates resource now requires the status updates feature to
be enabled. Otherwise, it will return a 403 forbidden.

Testing done:
Hit the status updates API with the feature enabled and disabled. It
returned a 403 when the feature was disabled and behaved correctly when
it was enabled.

Reviewed at https://reviews.reviewboard.org/r/8477/