Flesh out the front page of the Users Guide.
Our Review Board manuals have started out with just a plain Table of
Contents, with no introduction or real flow to the rest of the
documentation. This could be frustrating for new users.

This change is the first in a series that aims to modernize our
documentation, guiding users through getting going and pointing them to
more useful content without having to dig or search.

The new front page contains actual content, guiding the user through the
main pieces of the product, and grouping these pieces into digestable
sections.

The content on the pages linked from the main page haven't changed (with
the exception of some pages gaining an `:orphan:` tag to prevent Sphinx
errors when building). Edits will probably be coming to some of that
pages as I go.

Testing Done:
Built the docs and went through them. Didn't see any new errors.

Reviewed at https://reviews.reviewboard.org/r/8484/