Add the URL Avatar Service as a default service
The `URLAvatarService` is a new avatar service from Djblets to allow
absolute URLs for avatars. Users do not have access to this resource.
Instead, it is intended to be used with users that are created by
extensions so that they can provide their own avatars.

Testing Done:
- Ran unit tests.
- Saw that the new avatar service was enabled by default.
- Could not select the new avatar service in the config form.

Reviewed at https://reviews.reviewboard.org/r/8482/