Add previous and next buttons to navigate between file attachments
The file attachment UI previously did not have any way to navigate
between any file attachments. A user would have to navigate back to the
review request and select the next one they wished to review.

Now there are previous and next buttons on the edges of the page when
there are more file attachments to review. They follow the order of the
review request UI and only navigate to active (and latest revisions).
When they are hovered over, they expand out to show the thumbnail of the
attachment that will be navigated to.

Testing Done:
Ran unit tests.

Tested in the following browsers on OSX:

- Chrome
- Firefox release
- Firefox developer edition
- Safari

Verified the links were correct on local site review requests.

Reviewed at https://reviews.reviewboard.org/r/8443/