Fix a couple regressions for general comments.
The addition of a feature for general comments had a couple small issues. First
of all, we want to enable this feature by default, so that it's available for
people using release-3.0.x without any special configuration. Second, the
plumbing for the option through to the review dialog (to add the "Add Comment"
button) wasn't using the right name.

Testing done:
Was able to use general comments on my devserver again.

Reviewed at https://reviews.reviewboard.org/r/8502/