Flesh out the front page of the Administration Guide.
This change follows up on the Users Guide improvements, fleshing out the
content of the front page of the Administration Guide. This walks admins
through the installation and upgrade steps, and groups the various
settings and configuration steps needed to get started.

Future changes may flesh some of this out more, but for now, it's a much
better experience for new admins, and even old ones trying to locate
something they need to look up.

The plan later on is to add a new section on guides for common setups
(like setting up GitHub, Bitbucket, etc.).

Testing Done:
Built the docs and read through them.

Reviewed at https://reviews.reviewboard.org/r/8485/