Fix setting default avatar service
When RB was updated to use the new avatar services API, this was
mistakenly updated to look up the avatar service. However, the
`clean_avatar_service` method does this already, so we are attempting to
lookup an item with its ID equal to the class, which isn't going to
work. This would always fail, so submitting the form would unset the
default avatar service. Reverting this fixes the issue,

Testing Done:
Able to save the avatar services admin form and the default service gets
changed correctly.

Reviewed at https://reviews.reviewboard.org/r/8510/