Do not use BuiltinLocalsFieldMixin for the DiffCommitListField
The `DiffCommitListField` was using the `BuiltinLocalsFieldMixin` but
not taking advantage of it because it required the data in the actual
review request view as well as in the change description portions. Since
it was building all the data itself anyway, this class no longer uses
the mixin.

Testing Done:
It still works!

Reviewed at https://reviews.reviewboard.org/r/8542/