Fix general comment anchors
General comment anchors were subtly busted, missing a `>` to close the
anchor tag. This only resulted in broken rendering when a general review
comment was created *without* an issue. This has been fixed.

Testing Done:
Viewed a review containing a single general comment that did not raise
an issue. When the patch was not applied, the rendering issue appeared.
When the patch was not applied, the issue did not appear.

Reviewed at https://reviews.reviewboard.org/r/8541/