Fix up markup and styles for the review dialog.
We're going to be rewriting the review dialog to be very much improved, but in
the meantime, some layout and style problems had snuck in as a result of the
recent review changes. This change fixes those up as a stop-gap measure so that
we have something shippable before the new review experience work gets done. The
styles are mostly copied from the review entries, with appropriate changes to
make things fit in to the dialog. I haven't spent too much time making the code
perfectly factored because all of this is going to be rewritten anyway.

Testing done:
- Created a draft review with different types of comments and verified that
  everything looked good.
- Opened and closed editors for header, footer, and different comments, making
  sure that none of the text jumped around.

Reviewed at https://reviews.reviewboard.org/r/8545/