Fix header disappearing when clicking "Ship It!" in review dialog.
Some code in the review dialog was saving the model when the state of the "Ship
It!" checkbox was changed, which was causing some glitchy behavior with the
display of the header (when there were comments present). We already have code
to save the state of this checkbox when the dialog is being closed, so there's
no reason to do so every time it's clicked.

Testing done:
Added a diff comment, then opened the review dialog and clicked "Ship It".
Before this change, the header editor would close, potentially without saving
the data therein. After, it stays open.

Fixes bug 4414.

Reviewed at https://reviews.reviewboard.org/r/8546/