Add API docs for status updates.
This change adds the stub files and fixture data to generate API documentation
for the status updates resource. In doing this, I found that the examples for
this were showing 401 errors rather than the actual data, because of the new
`required_features` flag. I've added support for this to the API extension in a
couple ways. First, we temporarily enable the feature while doing the requests
to form the examples. Second, the required features are listed in the
documentation output.

Testing done:
Built HTML docs and verified the results.

Reviewed at https://reviews.reviewboard.org/r/8548/