Fix API tests which rely on required_features.
When an API resource depends on a feature, tests for it would pass or fail
depending on whether that feature was enabled in `settings_local.py`. Because
tests are there to test the actual behavior, we want to override the feature
status to enable it always.

This change does that, and fixes up the various mixins and top-level tests to
work correctly. I haven't added any extra tests to specific resources to verify
the API output when the feature(s) are disabled, because those are tested more
generically with a dummy resource and feature.

Testing done:
Disabled the status updates feature in my `settings_local.py` and ran unit
tests. Before this change, all the status updates tests failed. After, they
pass.

Reviewed at https://reviews.reviewboard.org/r/8550/