Add auto-complete functionality to the "Depends On" field
This change adds auto-complete to the "Depends On" field for review requests.
This allows people to not only search by ID but also by the summary string, and
shows a helpful drop-down which makes it easier to get the right review
requests.

Testing Done:
* Verified results using existing autocomplete tests.
* Added additional tests to verify new code:
    * Verify existence of autocomplete.
    * Verify autocomplete results.
* Addionally tested by hand:
    * Id and Summary tested for Depends On.

Reviewed at https://reviews.reviewboard.org/r/8434/