Fix loading of serialized comments for the image review UI.
The image review UI sends its serialized comments as an object rather than an
array, which wasn't at all clear from the code, and only emerged when I
actually tried to use it. I've changed the code back to using `_.each` to
iterate over it, and added a comment as to why.

Testing done:
Was able to use the image review UI again.

Reviewed at https://reviews.reviewboard.org/r/8573/