Support renamed files from GitHub in post-commit UI
The GitHub client did not support generating the diff for a file renamed
in a change that was retrieved via the post-commit review UI. This has
been corrected. Additionally, the code was somewhat cleaned up to not
shadow the `file` global. The unit test for `get_change` has been
updated to add a case for renamed files.

Testing Done:
Used the post-commit UI to create a new review request that contained a
renamed file. It worked as expected.

Reviewed at https://reviews.reviewboard.org/r/8577/