Fix expected number of queries in review request resource tests.
Change 6e19d96 on release-2.5.x added a link to the latest diff revision inside
the review request resource payload. That change contained a unit test to
ensure we didn't accidentally increase the number of queries, but the assumed
number wasn't correct when the change was merged to release-3.0.x. After
bisecting this and looking through everything, I've determined that the correct
course of action is indeed to just update the number of expected queries for
the 3.0 branch.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8580/